bill_detector/requirements.txt
Django==3.1.4
djangorestframework
opencv-python==4.5.3.56
pytesseract
Pillow
selenium==4.1.0
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
google-api-python-client
imgurpython
requests